requirements.txt
Flask
python-dotenv
line-bot-sdk
openai
